requirements.txt
django==5.1.1
djangorestframework==3.15.1
langchain==0.3.3
langchain-google-genai==2.0.1
python-dotenv==1.0.1
langchain-community